commands: doom: integrate with barebox
Compiling DOOM within barebox throws a lot of warnings.
Fix them and a new doomgeneric_barebox.c for barebox-specific
glue. Doomgeneric doesn't handle everything, it assumes a standard
C library, as we don't have one, sprinkle around some #ifdef
__BAREBOX__ to handle the differences. These can be made to
execute unconditionally once everything work.

+++ commands/doom/deh_str.h
@@ -18,8 +18,6 @@
 #ifndef DEH_STR_H
 #define DEH_STR_H
 
-#include <stdio.h>
-
 #include "doomfeatures.h"
 
 // Used to do dehacked text substitutions throughout the program
@@ -29,7 +27,7 @@
 char *DEH_String(char *s);
 void DEH_printf(char *fmt, ...);
 void DEH_fprintf(FILE *fstream, char *fmt, ...);
-void DEH_snprintf(char *buffer, size_t len, char *fmt, ...);
+void DEH_snprintf(char *buffer, unsigned len, char *fmt, ...);
 void DEH_AddStringReplacement(char *from_text, char *to_text);
